+++ yukon-server/COMMON/include/database_reader.h
@@ -47,7 +47,7 @@ public:
     RowReader &operator>>(UINT &operand);
     RowReader &operator>>(UCHAR &operand);
     RowReader &operator>>(unsigned long &operand);
-    RowReader &operator>>(__int64 &operand);
+    RowReader &operator>>(long long &operand);
     RowReader &operator>>(double &operand);
     RowReader &operator>>(float &operand);
     RowReader &operator>>(CtiTime &operand);
@@ -63,7 +63,7 @@ public:
     RowReader &operator<<(const INT operand);
     RowReader &operator<<(const UINT operand);
     RowReader &operator<<(const unsigned long operand);
-    RowReader &operator<<(const __int64 operand);
+    RowReader &operator<<(const long long operand);
     RowReader &operator<<(const double operand);
     RowReader &operator<<(const float operand);
     RowReader &operator<<(const CtiTime &operand);
