+++ yukon-server/COMMON/include/truncatingConsoleAppender.h
@@ -12,8 +12,8 @@ namespace Logging {
 
 class IM_EX_CTIBASE TruncatingConsoleAppender : public log4cxx::WriterAppender
 {
-    __int64 _interval;
-    __int64 _intervalEnd;
+    long long _interval;
+    long long _intervalEnd;
     unsigned _maxBurstSize, _currentBurst;
     static const size_t BurstBufferLength = 10;
     boost::circular_buffer<log4cxx::spi::LoggingEventPtr> _burstBuffer;
