+++ yukon-server/COMMON/include/win_helper.h
@@ -33,6 +33,6 @@ inline std::string getSystemErrorMessage(const DWORD errorCode)
     return result;
 }
 
-IM_EX_CTIBASE __int64 getPrivateBytes();
+IM_EX_CTIBASE long long getPrivateBytes();
 
 } // namespace Cti
